requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.1
django-heroku==0.2.0
djangorestframework==3.7.7
dj-database-url==0.4.2
flake8==3.5.0
gunicorn==19.7.1
idna==2.6
mccabe==0.6.1
nose==1.3.7
pew==1.1.2
pipenv==9.0.3
psycopg2==2.7.3.2
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
whitenoise==3.3.1
